backend/requirements.txt
google-ads
langchain==0.3.9
langchain-community==0.3.9
html2text
langchain_groq
fastapi
uvicorn
Pillow